chore: rename canister_build_config model
